Implement the moving cube in ggez

+++ src/models.rs
@@ -1,23 +1,23 @@
 pub trait Movable {
-    fn move_by(&mut self, x: f64, y: f64);
-    fn move_to(&mut self, x: f64, y: f64);
+    fn move_by(&mut self, x: f32, y: f32);
+    fn move_to(&mut self, x: f32, y: f32);
 }
 
 #[derive(Debug)]
 pub struct Position {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 #[derive(Debug)]
 pub struct Player {
-    pub width: usize,
-    pub height: usize,
+    pub width: f32,
+    pub height: f32,
     pub pos: Position,
 }
 
 impl Player {
-    pub fn new(x: f64, y: f64, size: usize) -> Self {
+    pub fn new(x: f32, y: f32, size: f32) -> Self {
         Self {
             width: size,
             height: size,
@@ -27,12 +27,12 @@ impl Player {
 }
 
 impl Movable for Player {
-    fn move_by(&mut self, x: f64, y: f64) {
+    fn move_by(&mut self, x: f32, y: f32) {
         self.pos.x += x;
         self.pos.y += y;
     }
 
-    fn move_to(&mut self, x: f64, y: f64) {
+    fn move_to(&mut self, x: f32, y: f32) {
         self.pos.x = x;
         self.pos.y = y;
     }
